Antrea deployment and custom logs task
This commit includes the changes made and the files obtained by
following the instruction of the common task mentioned at:
https://github.com/antrea-io/antrea/discussions/5976

Antrea agent and controller code was modified to include 2 log entries
to indicate the name of the pod in which the agents and controller are
running.

It also contains the version of Open vSwitch being used in the
`antrea-ovs` container of the `antrea-agent` pod.

The `antrea-version.txt` file shows the output of the `version` command
of the Antrea CLI binary downloaded from Github release assests for the
latest Antrea release.

Signed-off-by: VaibhavMalik4187 <vaibhavmalik2018@gmail.com>